Refactor SPI
