+++ make/mcu/STM32F1.mk
@@ -61,6 +61,7 @@ VCP_SRC = \
 MCU_COMMON_SRC = \
             drivers/adc_stm32f10x.c \
             drivers/bus_i2c_stm32f10x.c \
+            drivers/bus_std_periph.c \
             drivers/dma.c \
             drivers/inverter.c \
             drivers/light_ws2811strip_stdperiph.c \
